refactor: move file to sso folder
This is a sso related file
